infrastructure/api/requirements.txt
marshmallow~=3.20.1
environs~=9.0
jinja2~=3.1.5
uvicorn
fastapi
loguru
python-jose[cryptography]
passlib[bcrypt]

aiogram~=3.0

sqlalchemy~=2.0
alembic~=1.0
asyncpg

